// apps/web/src/libs/api/generated/models/resolveInterlineParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * SaaS API
 * NestJS + Fastify backend. Auth is served by Better Auth at /api/auth/* (not part of this spec).
 * OpenAPI spec version: 1.0.0
 */

export type ResolveInterlineParams = {
/**
 * @pattern ^[A-Z0-9]{2}$
 */
inboundAirline: string;
/**
 * @pattern ^[A-Z0-9]{2}$
 */
outboundAirline: string;
};